requirements.txt
torch
transformers
python-dotenv
SentencePiece
streamlit 
groq

fastapi
uvicorn[standard]
openai==0.28
faiss-cpu
langchain-google-genai
langchain
langchain-community
python-multipart
rank-bm25
sentence_transformers

networkx
python-docx
markdown
PyPDF2


#database packages
motor
pymongo[srv]


tree-sitter